refactor: move SafariButtonWarning to a separated file
